peerdas/pyproject.toml
[project]
name = "peerdas"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clickhouse-connect>=0.8.0",
    "pandas>=2.0",
    "plotly>=5.0",
    "altair>=5.0",
    "numpy>=1.26",
    "python-dotenv>=1.0",
    "nbconvert>=7.16.6",
    "pyarrow>=22.0.0",
    "kaleido>=1.2.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
